ci: enforce the presentation boundary with depguard (#1087)
Four rules, so a regression fails the build rather than review:
presentation and bubbletea are denied outside internal/presentation (cmd and
internal/container are composition roots and exempt), go-telegram is confined
to presentation/telegram, Playwright to browser/, robotgo to computer/.

MCPServerStatus and MCPServerStatusUpdateEvent move to agent/domain, next to
the MCPManager port that documents them. That was the last import dragging
internal/services into the presentation graph, so no exemption is needed.
