+++ .golangci.yml
@@ -17,6 +17,7 @@ linters:
     - gocognit
     - nestif
     - loggercheck
+    - depguard
   disable:
     - exhaustruct
     - varnamelen
@@ -24,6 +25,44 @@ linters:
     - paralleltest
     - testpackage
   settings:
+    depguard:
+      rules:
+        presentation-is-a-leaf:
+          list-mode: lax
+          files:
+            - "**/internal/**"
+            - "!**/internal/presentation/**"
+            - "!**/internal/container/**"
+            - "!$test"
+          deny:
+            - pkg: github.com/inference-gateway/cli/internal/presentation
+              desc: "domain, capability and platform contexts must not depend on a UI surface (#1087); internal/container is the composition root and is exempt"
+            - pkg: charm.land/bubbletea
+              desc: "bubbletea belongs to internal/presentation and cmd only (#1087)"
+        telegram-only-in-presentation:
+          list-mode: lax
+          files:
+            - "**"
+            - "!**/internal/presentation/telegram/**"
+          deny:
+            - pkg: github.com/go-telegram/bot
+              desc: "the Telegram driver lives in internal/presentation/telegram; consume internal/channels instead (#1087)"
+        playwright-only-in-browser:
+          list-mode: lax
+          files:
+            - "**"
+            - "!**/internal/browser/**"
+          deny:
+            - pkg: github.com/mxschmitt/playwright-go
+              desc: "Playwright is confined to the browser capability (#1087)"
+        robotgo-only-in-computer:
+          list-mode: lax
+          files:
+            - "**"
+            - "!**/internal/computer/**"
+          deny:
+            - pkg: github.com/go-vgo/robotgo
+              desc: "robotgo is confined to the computer capability (#1087)"
     gocyclo:
       min-complexity: 25
     cyclop:
